requirements.txt
discord.py
pytz
python-dotenv
mysql-connector-python==8.0.22
